后台datatable

+++ app/Http/Controllers/Backend/Admin/CampusController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers\Backend\Admin;
 
 use Illuminate\Http\Request;
@@ -8,19 +7,25 @@ use App\Http\Requests\CampusRequest;
 use App\Http\Controllers\Controller;
 use App\Unis\School\Campus;
 use App\Unis\School\School;
+use App\DataTables\CampusDataTable;
 
 class CampusController extends Controller
 {
-    public function index()
+    // public function index()
+    // {
+    // 	$campuses = Campus::with('school')->orderBy('school_id', 'desc')->paginate(config('site.perPage'));
+    // 	return view('backend.admin.campus.index', compact('campuses'));
+    // }
+
+    public function index(CampusDataTable $dataTable)
     {
-    	$campuses = Campus::with('school')->orderBy('school_id', 'desc')->paginate(config('site.perPage'));
-    	return view('backend.campus.index', compact('campuses'));
+        return $dataTable->render('backend.admin.campus.index');
     }
 
     public function create(Request $request)
     {
     	$school = School::select('id', 'name')->find($request->input('school_id'));
-    	return view('backend.campus.create', compact('school'));
+    	return view('backend.admin.campus.create', compact('school'));
     }
 
     public function store(CampusRequest $request)
@@ -32,13 +37,13 @@ class CampusController extends Controller
     public function show($campus)
     {
     	$campus = Campus::with('school')->find($campus);
-    	return view('backend.campus.show', compact('campus'));
+    	return view('backend.admin.campus.show', compact('campus'));
     }
 
     public function edit($campus)
     {
     	$campus = Campus::with('school')->find($campus);
-    	return view('backend.campus.edit', compact('campus'));
+    	return view('backend.admin.campus.edit', compact('campus'));
     }
 
     public function update(CampusRequest $request, Campus $campus)
@@ -47,9 +52,18 @@ class CampusController extends Controller
     	return redirect('/admin/campus')->with('success', '更新成功！');
     }
 
-    public function delete(Request $request)
+    // public function delete(Request $request)
+    // {
+    // 	$campus->delete();
+    // 	return redirect()->back()->with('success', '删除成功！');
+    // }
+
+    public function destroy(Campus $campus)
     {
-    	$campus->delete();
-    	return redirect()->back()->with('success', '删除成功！');
+        if ($campus->dorms->count() > 0 || $campus->canteens->count() > 0){
+            return redirect()->back()->with('failed', '还有食堂|宿舍存在，不能删除！');
+        }
+        $campus->delete();
+        return redirect()->back()->with('success', '删除成功！');
     }
 }
